Navigation start setup ok, now with back and home and refactored navigation and layout to separate files
