+++ app/src/main/java/com/example/stockpalapp/ui/screens/HomeScreen.kt
@@ -14,9 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavController
-import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
 import com.example.stockpalapp.AppLayout
 import com.example.stockpalapp.R
 import com.example.stockpalapp.ui.theme.StockPalAppTheme
@@ -72,7 +70,8 @@ fun HomeScreen(navController: NavController){
         actionIcon = null,
         navigationContentDescription = null,
         actionContentDescription = null,
-        navController = navController
+        navController = navController,
+        leftIconClickHandler = {}
     )
 }
